% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketable_list.R
\name{bucket_list}
\alias{bucket_list}
\title{Create a bucket list.}
\usage{
bucket_list(header = NULL, ..., group_name, group_put_max = rep(Inf,
  length(labels)), selector = NULL, options = sortable_options(),
  style = css_bucket_list())
}
\arguments{
\item{header}{Text that appears at the top of the bucket list.  (This is
encoded as an HTML \code{<p>} tag, so not strictly speaking a header.)}

\item{...}{One or more specifications for a rank list, and must be
defined by \link{add_rank_list}.}

\item{group_name}{Passed to \code{sortable.js} as the group name}

\item{group_put_max}{Not yet implemented}

\item{selector}{This is the css id to use, and must be unique in your shiny
app. If NULL, the function generates a selector of the form
\code{rank_list_id_1}, and will automatically increment for every
\code{rank_list}.}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{style}{A css stylesheet, provided as a character string.  See also \code{\link[=css_rank_list]{css_rank_list()}}.}
}
\description{
A bucket list can contain more than one \link{rank_list} and allows
drag-and-drop of items between the different lists.
}
\examples{
## Example of a shiny app
if (interactive()) {
  app <- system.file("shiny-examples/bucket_list_app.R", package = "sortable")
  shiny::runApp(app)
}
## -- example-bucket-list ---------------------------------------------

## bucket list

bucket_list(
  header = "This is a bucket list. You can drag items between the lists.",
  group_name = "bucketable",
  add_rank_list(
    text = "Drag from here",
    labels = c("a", "bb", "ccc"),
    input_id = NULL
  ),
  add_rank_list(
    text = "to here",
    labels = NULL,
    input_id = NULL
  )
)

## bucket list with three columns

bucket_list(
  header = c("Sort these items into Letters and Numbers"),
  add_rank_list(
    text = "Drag from here",
    labels = sample(c(1:3, letters[1:2])),
    input_id = NULL
  ),
  add_rank_list(
    text = "Letters",
    input_id = NULL
  ),
  add_rank_list(
    text = "Numbers",
    input_id = NULL
  )
)


}
