% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_bucket_list}
\alias{css_bucket_list}
\alias{css_parsons}
\title{CSS stylesheet for a bucket list.}
\usage{
css_bucket_list()

css_parsons()
}
\description{
This is the default CSS stylesheet for the \link{bucket_list} function.
}
