% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_bucketable_list}
\alias{css_bucketable_list}
\alias{css_parsons}
\title{CSS stylesheet for a bucketable list.}
\usage{
css_bucketable_list()

css_parsons()
}
\description{
This is the default CSS stylesheet for the \link{bucketable_list} function.
}
