% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_parsons_problem}
\alias{css_parsons_problem}
\title{CSS stylesheet for sortable list.}
\usage{
css_parsons_problem()
}
\description{
This is the default CSS stylesheet for the \link{sortable_list} function.
}
