% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{is_modules_enabled}
\alias{is_modules_enabled}
\alias{enable_modules}
\title{Check if shiny modules are enabled for \code{sortable}.}
\usage{
is_modules_enabled()

enable_modules(enable = TRUE)
}
\arguments{
\item{enable}{If \code{TRUE}, enables modules. If \code{FALSE} disables them.}
}
\value{
Logical value indicating whether shiny modules are enabled or not.
}
\description{
Due to an early (regrettable) design decision, \code{sortable} in versions <= 0.5.0
did not support shiny modules.
To use \code{sortable} with shiny modules, you have to opt in to different
behaviour, by calling \code{enable_modules()}.
}
