% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsons.R
\name{parsons}
\alias{parsons}
\title{Create a parsons problem.}
\usage{
parsons(header = NULL, text = c("Drag from here",
  "Construct your solution here"), labels, input_id, group_name,
  options = sortable_options(), style = css_parsons())
}
\arguments{
\item{header}{Text that appears at the top of the bucketable list.  (This is
encoded as an HTML \code{<p>} tag, so not strictly speaking a header.)}

\item{text}{Vector of headings for each column.}

\item{labels}{List of \code{label} to pass (individually) to \code{\link[=sortable_list]{sortable_list()}}.}

\item{input_id}{Character vector of \code{input_id} to pass (individually) to \code{\link[=sortable_list]{sortable_list()}}.}

\item{group_name}{Passed to \code{sortable.js} as the group name}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{style}{A css stylesheet, provided as a character string.  See also \code{\link[=css_sortable_list]{css_sortable_list()}}.}
}
\description{
Create a parsons problem.
}
\examples{
## Example of a shiny app
if (interactive()) {
  app <- system.file("shiny-examples/parsons_app.R", package = "sortable")
  shiny::runApp(app)
}
## -- example-parsons -----------------------------------------------------

## Parson's problem

parsons(
  header = "This is an example of a Parsons problem",
  text = c("Drag from here", "Construct your solution here"),
  labels = c(
      "iris",
      "mutate(...)",
      "summarize(...)",
      "print()"
  ),
  input_id = "input_parsons",
  group_name = "parsons_test"
)


parsons(
  labels = c(
    "iris",
    "mutate(...)",
    "summarize(...)",
    "print()"
  ),
  input_id = "input_parsons",
  group_name = "parsons_test"
)
}
