% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_rank.R
\name{question_is_correct.rank}
\alias{question_is_correct.rank}
\title{#' @export
question_is_valid.rank <- function(question, answer_input, ...) {
!is.null(answer_input)
}}
\usage{
\method{question_is_correct}{rank}(question, answer_input, ...)
}
\description{
#' @export
question_is_valid.rank <- function(question, answer_input, ...) {
!is.null(answer_input)
}
}
