% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_js.R
\name{render_sortable}
\alias{render_sortable}
\title{Widget render function for use in Shiny.}
\usage{
render_sortable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This is useful
if you want to save an expression in a variable.}
}
\description{
Widget render function for use in Shiny.
}
