% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable-package.R
\docType{package}
\name{sortable-package}
\alias{sortable-package}
\alias{_PACKAGE}
\title{sortable: Create Drag-and-Drop Elements Using the JavaScript `sortable.js` Library}
\description{
Make almost anything sortable, by adding drag-and-drop behavior to 
    Shiny Apps and gadgets.  Add ranking, sorting and bucketing question types
    to `learnr` tutorials. The package creates an 'htmlwidget' by wrapping 
    the 'sortable.js' JavaScript library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/sortable}
  \item Report bugs at \url{https://github.com/rstudio/sortable/issues}
}

}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com}

Authors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com}
  \item Kenton Russell \email{kent.russell@timelyportfolio.com} (Original author) [conceptor]
}

Other contributors:
\itemize{
  \item Lebedev Konstantin ('Sortable.js' library,
        http://sortablejs.github.io/Sortable/) [copyright holder]
}

}
\keyword{internal}
