% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{question_initialize_input.sortable}
\alias{question_initialize_input.sortable}
\alias{question_completed_input.sortable}
\alias{question_is_valid.sortable}
\alias{question_is_correct.sortable}
\title{Learnr methods for \link{sortable_js}}
\usage{
\method{question_initialize_input}{sortable}(question, answer_input, ...)

\method{question_completed_input}{sortable}(question, answer_input, ...)

\method{question_is_valid}{sortable}(question, answer_input, ...)

\method{question_is_correct}{sortable}(question, answer_input, ...)
}
\arguments{
\item{question}{question object used}

\item{answer_input}{input value provided by `input$answer`}

\item{...}{future parameter expansion and custom arguments to be used in dispatched s3 methods.}
}
\description{
Questions will be created using the first set of options.
Each set of answer options must contain the same set of answer options.
When the question is completed, the first correct answer will be displayed.
}
\examples{
if (require(learnr, quietly = TRUE)) {
  # to be used within a learnr tutorial...
  question_sortable(
    "Sort the first 5 letters",
    answer(letters[1:5], correct = TRUE),
    allow_retry = TRUE
  )

}
}
\seealso{
\code{\link{question_sortable}}
}
