% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_js.R
\name{sortable_output}
\alias{sortable_output}
\title{Widget output function for use in Shiny.}
\usage{
sortable_output(input_id, width = "0px", height = "0px")
}
\arguments{
\item{input_id}{output variable to use for the sortable object}

\item{width}{\code{integer} width in pixels.  Since the container is expected
to not be displayed, \code{0} is the default.}

\item{height}{\code{integer} height in pixels.  Since the container is expected
to not be displayed, \code{0} is the default.}
}
\description{
Widget output function for use in Shiny.
}
