% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{sortable_question}
\alias{sortable_question}
\title{Learnr methods for sortable_learnr}
\usage{
\method{question_initialize_input}{sortable}(...,
  options = sortable_options())
}
\arguments{
\item{...}{parameters passed onto \code{learnr::\link[learnr]{question}}.}
}
\description{
Questions will be created using the first set of options.
Each set of answer options must contain the same set of answer options.
When the question is completed, the first correct answer will be displayed.
}
\examples{
if (require(learnr, quietly = TRUE)) {
  # to be used within a learnr tutorial...
  sortable_question(
    "Sort the first 5 letters",
    answer(letters[1:5], correct = TRUE),
    type = "sortable",
    allow_retry = TRUE,
    options = sortable_options(animation = 150)
  )
}
}
