% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortableItemlist.R
\name{sortableItemlist}
\alias{sortableItemlist}
\title{Sortable item list.}
\usage{
sortableItemlist(outputId, labels, selector, class = "list-group-item")
}
\arguments{
\item{class}{}
}
\description{
Sortable item list.
}
