% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable.R
\name{sortableOutput}
\alias{sortableOutput}
\title{Widget output functionA for use in Shiny.}
\usage{
sortableOutput(outputId, width = "0px", height = "0px")
}
\arguments{
\item{outputId}{output variable to use for the sortable object}

\item{width}{\code{integer} width in pixels.  Since the container is expected
to not be displayed, \code{0} is the default.}

\item{height}{\code{integer} height in pixels.  Since the container is expected
to not be displayed, \code{0} is the default.}
}
\description{
Widget output functionA for use in Shiny.
}
