% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sortableR.R
\name{sortableR}
\alias{sortableR}
\title{Sortable.js for R}
\usage{
sortableR(message, width = NULL, height = NULL)
}
\description{
htmlwidget for \href{http://rubaxa.github.io/Sortable/}{Sortable.js}
to use in interactive HTML in the Viewer, Browser, or Shiny.
}

