% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{update_bucket_list}
\alias{update_bucket_list}
\title{Change the value of a bucket list.}
\usage{
update_bucket_list(
  input_id,
  text = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
At the moment, you can only update the \code{text} of the \code{bucket_list}, not the
labels.
}
\seealso{
\link{bucket_list}
}
