% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{update_rank_list}
\alias{update_rank_list}
\title{Change the value of a rank list on the client.}
\usage{
update_rank_list(
  input_id,
  text = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{input_id}{output variable to read the plot/image from.}

\item{text}{Text to appear at top of list.}

\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}
}
\description{
Change the value of a rank list on the client.
}
