% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketable_list.R
\name{bucketable_list}
\alias{bucketable_list}
\title{Create a bucketable list.}
\usage{
bucketable_list(header, group_name, ..., group_put_max = rep(Inf,
  length(labels)), selector = NULL, options = sortable_options(),
  style = css_bucketable_list())
}
\arguments{
\item{header}{Text that appears at the top of the bucketable list.  (This is
encoded as an HTML \code{<p>} tag, so not strictly speaking a header.)}

\item{group_name}{Passed to \code{sortable.js} as the group name}

\item{...}{One or more specifications for a sortable list, and must be
defined by \link{add_sortable_list}.}

\item{group_put_max}{Not yet implemented}

\item{selector}{This is the css id to use, and must be unique in your shiny
app. If NULL, the function generates a selector of the form
\code{sortable_list_id_1}, and will automatically increment for every
\code{sortable_list}.}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{style}{A css stylesheet, provided as a character string.  See also \code{\link[=css_sortable_list]{css_sortable_list()}}.}
}
\description{
A bucketable list can contain more than one \link{sortable_list} and allows
drag-and-drop of items between the different lists.
}
\examples{
## Example of a shiny app
if (interactive()) {
  app <- system.file("shiny-examples/bucketable_list_app.R", package = "sortable")
  shiny::runApp(app)
}
## -- example-bucketable-list ---------------------------------------------

## Bucketable list

bucketable_list(
  header = "This is a bucketable list. You can drag items between the lists.",
  group_name = "bucketable",
  add_sortable_list(
    text = "Drag from here",
    labels = c("a", "bb", "ccc"),
    input_id = NULL
  ),
  add_sortable_list(
    text = "to here",
    labels = NULL,
    input_id = NULL
  )
)

## Bucketable list with three columns

bucketable_list(
  header = c("Sort these items into Letters and Numbers"),
  group_name = "bucket_test",
  add_sortable_list(
    text = "Drag from here",
    labels = sample(c(1:3, letters[1:2])),
    input_id = NULL
  ),
  add_sortable_list(
    text = "Letters",
    input_id = NULL
  ),
  add_sortable_list(
    text = "Numbers",
    input_id = NULL
  )
)


}
