% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketable_list.R
\name{bucketable_list}
\alias{bucketable_list}
\title{Create a bucketable item list.}
\usage{
bucketable_list(output_ids, labels_list, headings, group_name,
  group_put_max = rep(Inf, length(labels)), selector = NULL,
  options = sortable_options())
}
\arguments{
\item{output_ids}{Character vector of \code{output_id} to pass (individually) to \link{sortable_list}}

\item{labels_list}{List of \code{label} to pass (individually) to \link{sortable_list}}

\item{headings}{Vector of headings for each column}

\item{group_name}{Passed to \code{sortable.js} as the group name}

\item{group_put_max}{Not yet implemented}

\item{selector}{This is the css id to use, and must be unique in your shiny
app. If NULL, the function generates a selector of the form
\code{sortable_list_id_1}, and will automatically increment for every
\code{sortable_list}.}

\item{options}{Options to be supplied to Sortable. See \link{sortable_options} for more details}
}
\description{
Create a bucketable item list.
}
