% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketable_list.R
\name{bucketable_list}
\alias{bucketable_list}
\title{Create a bucketable item list.}
\usage{
bucketable_list(output_ids, labels_list, group_name,
  group_put_max = rep(Inf, length(labels)), selector = NULL,
  class = "list-group-item", options = sortable_options())
}
\arguments{
\item{selector}{This is the css id to use, and must be unique in your shiny
app. If NULL, the function generates a selector of the form
\code{sortable_list_id_1}, and will automatically increment for every
\code{sortable_list}.}

\item{class}{The css class to use}

\item{options}{Options to be supplied to Sortable. See \link{sortable_options} for more details}
}
\description{
Create a bucketable item list.
}
\examples{
# example-sortable-list ---------------------------------------------------

if (require(htmltools)) {

  list_group_css <- '
  .list-group-item {
    position: relative;
    display: block;
    padding: 10px 15px;
    margin-bottom: -1px;
    background-color: #fff;
      border: 1px solid #ddd;
  }'

  html_print(
    tagList(
      tags$style(list_group_css),
      tags$p("You can drag and drop these list items:"),
      sortable_list("example_2", labels = c("one", "two", "three"))
    )
  )
}
}
\seealso{
\link{sortable}
}
