% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_options.R
\name{is_sortable_options}
\alias{is_sortable_options}
\title{Check if object is sortable options.}
\usage{
is_sortable_options(x)
}
\arguments{
\item{x}{Object to test}
}
\description{
Check if object is sortable options.
}
