% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_rank.R
\name{question_rank}
\alias{question_rank}
\title{Ranking question for learnr tutorials.}
\usage{
question_rank(..., random_answer_order = TRUE,
  options = sortable_options())
}
\arguments{
\item{...}{parameters passed onto \code{learnr::\link[learnr]{question}}.}

\item{random_answer_order}{Display answers in a random order.}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}
}
\description{
Add interactive ranking tasks to your \code{learnr} tutorials.  The student can
drag-and-drop the answer options into the desired order.
}
\details{
Each set of answer options must contain the same set of answer options. When
the question is completed, the first correct answer will be displayed.

Note that, by default, the answer order is randomized.
}
\examples{
if (require(learnr, quietly = TRUE)) {
  # to be used within a learnr tutorial...
  question_rank(
    "Sort the first 5 letters",
    answer(letters[1:5], correct = TRUE),
    allow_retry = TRUE,
    options = sortable_options(animation = 150)
  )
}
}
