% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{rank_list}
\alias{rank_list}
\title{Create a ranking item list.}
\usage{
rank_list(text = "", labels, input_id, selector = NULL,
  options = sortable_options(), additional_class = "",
  style = css_rank_list())
}
\arguments{
\item{text}{Text to appear at top of list.}

\item{labels}{A character vector with the text to display inside the widget.}

\item{input_id}{output variable to read the plot/image from.}

\item{selector}{This is the css id to use, and must be unique in your shiny
app. If NULL, the function generates a selector of the form
\code{rank_list_id_1}, and will automatically increment for every
\code{rank_list}.}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{additional_class}{Additional css class name to use. This gets appended to the \code{rank-list} class, and is used by the \code{\link[=bucket_list]{bucket_list()}} function.}

\item{style}{A css stylesheet, provided as a character string.  See also \code{\link[=css_rank_list]{css_rank_list()}}.}
}
\description{
Creates a ranking item list using the \code{sortable.js} framework, and generates
an \code{htmlwidgets} element.  The elements of this list can be dragged and
dropped in any order.  Typically you will embed a \code{rank_list} inside a
Shiny app or any shiny runtime, e.g. a \code{learnr} tutorial.
}
\details{
The widget automatically updates a Shiny output, with the matching \code{input_id}.
}
\examples{
## Example of a shiny app
if (interactive()) {
  app <- system.file("shiny-examples/rank_list_app.R", package = "sortable")
  shiny::runApp(app)
}
## - example-rank-list ------------------------------------------------

rank_list(
  text = "You can drag, drap and re-order these items:",
  labels = c("one", "two", "three", "four", "five"),
  input_id = "example_2"
)


}
\seealso{
\link{sortable_js}, \link{bucket_list}, \link{parsons}
}
