% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sortableR.R
\name{renderSortableR}
\alias{renderSortableR}
\title{Widget render function for use in Shiny}
\usage{
renderSortableR(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

