% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable.R
\name{sortable}
\alias{sortable}
\title{Sortable.js for R}
\usage{
sortable(selector, options = NULL, width = 0, height = 0)
}
\arguments{
\item{selector}{\code{String} selector id on which to apply Sortable.  Note, Sortable
works with any html element, not just \code{ul/li}.}

\item{options}{\code{list} configuration options for Sortable.  For a throrough
discussion, see \href{https://github.com/SortableJS/Sortable#options}{Sortable Options}.
This is the list of supported options:
\itemize{
\item \code{group}
\item \code{sort}
\item \code{disabled}
\item \code{store}
\item \code{animation}
\item \code{handle}
\item \code{filter}
\item \code{draggable}
\item \code{ghostClass}
\item \code{scroll}
\item \code{scrollSensitivity}
\item \code{scrollSpeed}
\item \code{setData}
\item \code{onStart}
\item \code{onEnd}
\item \code{onAdd}
\item \code{onUpdate}
\item \code{onSort}
\item \code{onRemove}
\item \code{onFilter}
}}

\item{width}{\code{integer} width in pixels.  Since the container is expected
to not be displayed, \code{0} is the default.}

\item{height}{\code{integer} height in pixels.  Since the container is expected
to not be displayed, \code{0} is the default.}
}
\description{
htmlwidget that provides \href{https://github.com/SortableJS/Sortable}{Sortable.js}
to use for additional interactivity in the Viewer, Browser, or Shiny.
}
\examples{
# Simple example of sortable. Set the tags `id` equal to the sortable `selector`
if (require(htmltools)) {
  html_print(
    tagList(
      tags$ul(
        id = "example02",
        tags$li(tags$h4("drag me")),
        tags$li(tags$h4("sort me")),
        tags$li(tags$h4("touch me")),
        tags$li(tags$h4("any way you like"))
      ),
      sortable(selector = "example02")
    )
  )
}
}
