% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable-package.R
\docType{package}
\name{sortable}
\alias{sortable}
\alias{_PACKAGE}
\alias{sortable-package}
\title{sortable: Create Drag-and-Drop Elements Using the JavaScript 'SortableJS' Library}
\description{
Add drag-and-drop behavior to Shiny Apps and gadgets.
    Add ranking, sorting and bucketing question types to `learnr` tutorials.
    The package creates an 'htmlwidget' by wrapping the 'SortableJS'
    JavaScript library.
}
\section{A new html widget}{

\itemize{
\item \code{\link[=sortable_js]{sortable_js()}} is a low-level function that adds the \code{SortableJS} to your widgets.
}
}

\section{Important functions}{


The important functions in this package are:
\itemize{
\item \code{\link[=rank_list]{rank_list()}} creates a drag-and-drop, rank list
\item \code{\link[=bucket_list]{bucket_list()}} lets you add multiple \code{rank_list} objects in columns
}
}

\section{Custom question types for \code{learnr}}{


You can also use new question types in your \code{learnr} tutorials:
\itemize{
\item \code{\link[=question_rank]{question_rank()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/sortable}
  \item Report bugs at \url{https://github.com/rstudio/sortable/issues}
}

}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com}

Authors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com}
  \item Kenton Russell \email{kent.russell@timelyportfolio.com} (Original author) [conceptor]
}

Other contributors:
\itemize{
  \item Lebedev Konstantin ('SortableJS' library,
        http://sortablejs.github.io/Sortable/) [copyright holder]
}

}
\keyword{internal}
