% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable-package.R
\docType{package}
\name{sortable}
\alias{sortable}
\alias{_PACKAGE}
\alias{sortable-package}
\title{sortable: Create Drag-and-Drop Elements Using the JavaScript `sortable.js` Library}
\description{
Add drag-and-drop behavior to Shiny Apps and gadgets.  
    Add ranking, sorting and bucketing question types to `learnr` tutorials. 
    The package creates an 'htmlwidget' by wrapping the 'sortable.js' 
    JavaScript library.
}
\section{A new html widget}{

\itemize{
\item \code{\link[=sortable_js]{sortable_js()}} is a low-level function that adds the \code{sortable.js} to your widgets.
}
}

\section{Important functions}{


The important functions in this package are:
\itemize{
\item \code{\link[=sortable_list]{sortable_list()}} creates a drag-and-drop, sortable list
\item \code{\link[=bucketable_list]{bucketable_list()}} lets you add multiple \code{sortable_list} objects in columns
\item \code{\link[=parsons]{parsons()}} sets up a Parson's problem, a useful question type when teaching code to students
}
}

\section{Custom question types for \code{learnr}}{


You can also use new question types in your \code{learnr} tutorials:
\itemize{
\item \code{\link[=question_sortable]{question_sortable()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/sortable}
  \item Report bugs at \url{https://github.com/rstudio/sortable/issues}
}

}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com}

Authors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com}
  \item Kenton Russell \email{kent.russell@timelyportfolio.com} (Original author) [conceptor]
}

Other contributors:
\itemize{
  \item Lebedev Konstantin ('Sortable.js' library,
        http://sortablejs.github.io/Sortable/) [copyright holder]
}

}
\keyword{internal}
