% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_js_capture_input.R
\name{sortable_js_capture_input}
\alias{sortable_js_capture_input}
\title{Construct JavaScript method to capture Shiny inputs on change.}
\usage{
sortable_js_capture_input(input_id)
}
\arguments{
\item{input_id}{The output id.}
}
\description{
This captures the inputs of a \code{sortable} list.  Typically you would use this
with the \code{onSort} option of \code{sortable_js}. See \code{\link[=sortable_options]{sortable_options()}}.
}
\examples{
# For an example, see the Shiny app at
system.file("shiny-examples/drag_vars_to_plot/drag_vars_to_plot_app.R", package = "sortable")
}
\seealso{
\link{sortable_js} and \link{rank_list}.
}
