% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{sortable_js_capture_input}
\alias{sortable_js_capture_input}
\alias{sortable_js_capture_bucket_input}
\title{Construct JavaScript method to capture Shiny inputs on change.}
\usage{
sortable_js_capture_input(input_id,
  child_callback = "function(child){return child.innerText}")

sortable_js_capture_bucket_input(input_id, input_ids, selectors)
}
\arguments{
\item{input_id}{Shiny input name to set}

\item{child_callback}{optional. A JavaScript callback function that executes on each item within the sortable. The default function returns item.innerText.}

\item{input_ids}{Set of Shiny input ids to set corresponding to the provided \code{selectors}}

\item{selectors}{Set of SortableJS selector values to help retrieve all to set as an object}
}
\description{
This captures the state of a \code{sortable} list.  Typically you would use this
with the \code{onSort} option of \code{sortable_js}. See \code{\link[=sortable_options]{sortable_options()}}.
}
\examples{
# For an example, see the Shiny app at
system.file("shiny-examples/drag_vars_to_plot/drag_vars_to_plot_app.R", package = "sortable")
}
\seealso{
\link{sortable_js} and \link{rank_list}.
}
