% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_list.R
\name{sortable_js_capture_input}
\alias{sortable_js_capture_input}
\title{Construct JS method to capture inputs on change.}
\usage{
sortable_js_capture_input(output_id)
}
\arguments{
\item{output_id}{The output id.}
}
\description{
This captures the inputs of a \code{sortable} list.  Typically you would use this
with the \code{onSort} option of \code{sortable}. See \code{\link[=sortable_options]{sortable_options()}}.
}
\examples{
# For an example, see the Shiny app at
system.file("shiny-examples/drag_vars_to_plot/app.R", package = "sortable")
}
\seealso{
\link{sortable} and \link{sortable_list}.
}
