% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_list.R
\name{sortable_list}
\alias{sortable_list}
\title{Sortable item list.}
\usage{
sortable_list(output_id, labels, selector = NULL,
  class = "list-group-item", options = sortable_options())
}
\arguments{
\item{output_id}{output variable to read the plot/image from.}

\item{labels}{A character vector with text to display.}

\item{selector}{This is the css id to use, and must be unique in your shiny
app. If NULL, the function generates a selector of the form
\code{sortable_list_id_1}, and will automatically increment for every
\code{sortable_list}}

\item{class}{The css class to use}

\item{options}{Options to be supplied to Sortable. See \link{sortable_options} for more details}
}
\description{
Creates a sortable item list using the \code{sortable.js} framework, and generates
an \strong{htmlwidgets} element.
}
\examples{
library(shiny)
library(sortable)

shuffle <- function(x)sample(x, size = length(x))

ui <- shinyUI(fluidPage(

  fluidRow(
    column(
      width = 12,
      tags$h2("This is a sortable list"),
      tags$p("Drag the items in the correct order"),
      sortableItemlist(
        "sortable_list_1",
        labels = shuffle(c("one", "two", "three", "four", "five"))
      )
    )
  ),
  fluidRow(
    verbatimTextOutput("results_list_1")
  )
))

server <-  function(input,output){
  output$results_list_1 <-
    renderPrint(
      input$sortable_list_1
    )

}


shiny::shinyApp(ui, server = server)
}
\seealso{
\link{sortable}
}
