% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_js.R
\name{sortable_output}
\alias{sortable_output}
\title{Widget output function for use in Shiny.}
\usage{
sortable_output(input_id, width = "0px", height = "0px")
}
\arguments{
\item{input_id}{output variable to use for the sortable object}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}
}
\description{
Widget output function for use in Shiny.
}
