% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sortableR.R
\name{sortableROutput}
\alias{sortableROutput}
\title{Widget output function for use in Shiny}
\usage{
sortableROutput(outputId, width = "0px", height = "0px")
}
\description{
Widget output function for use in Shiny
}

