% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{update_bucket_list}
\alias{update_bucket_list}
\title{Change the value of a bucket list.}
\usage{
update_bucket_list(
  css_id,
  header = NULL,
  labels = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{css_id}{This is the css id to use, and must be unique in your shiny
app. This defaults to the value of \code{group_id}, and will be appended to the
value "bucket-list-container", to ensure the CSS id is unique for the
container as well as the embedded rank lists.}

\item{header}{Text that appears at the top of the bucket list.  (This is
encoded as an HTML \verb{<p>} tag, so not strictly speaking a header.)  Note
that you must explicitly provide \code{header} argument, especially in the case
where you want the header to be empty - to do this use \code{header = NULL} or
\code{header = NA}.}

\item{labels}{A character vector with the text to display inside the widget.
This can also be a list of html tag elements.  The text content of each
label or label name will be used to set the shiny \code{input_id} value.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
At the moment, you can only update the \code{text} of the \code{bucket_list}, not the
labels.
}
\examples{
## Example of a shiny app that updates a bucket list and rank list
if (interactive()) {
  app <- system.file(
    "shiny-examples/update/app.R",
    package = "sortable"
  )
  shiny::runApp(app)
}
}
\seealso{
\link{bucket_list}
}
