% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{update_rank_list}
\alias{update_rank_list}
\title{Change the value of a rank list.}
\usage{
update_rank_list(
  input_id,
  text = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{input_id}{output variable to read the plot/image from.}

\item{text}{Text to appear at top of list.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}
}
\description{
At the moment, you can only update the \code{text} of the \code{rank_list}, not the
labels.
}
\seealso{
\link{rank_list}
}
