% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_list.R
\name{update_rank_list}
\alias{update_rank_list}
\title{Change the value of a rank list on the client.}
\usage{
update_rank_list(session = getDefaultReactiveDomain(), inputId, label = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}
}
\description{
Change the value of a rank list on the client.
}
